+++ src/middleware/packages/auth/index.js
@@ -1,5 +1,6 @@
 module.exports = {
-  AuthService: require('./services/auth'),
+  AuthLocalService: require('./services/auth.local'),
+  AuthOIDCService: require('./services/auth.oidc'),
   AuthAccountService: require('./services/account'),
   AuthJWTService: require('./services/jwt'),
 };
