+++ src/middleware/packages/auth/middlewares/localLogout.js
@@ -1,0 +1,6 @@
+const localLogout = (req, res, next) => {
+  req.logout(); // Passport logout
+  next();
+}
+
+module.exports = localLogout;
