+++ src/middleware/packages/auth/middlewares/redirectToFront.js
@@ -1,0 +1,14 @@
+const redirectToFront = (req, res) => {
+  // Redirect browser to the redirect URL pushed in session
+  let redirectUrl = new URL(req.session.redirectUrl);
+  if (req.user) {
+    // If a token was stored, add it to the URL so that the client may use it
+    if (req.user.token) redirectUrl.searchParams.set('token', req.user.token);
+    redirectUrl.searchParams.set('new', req.user.newUser ? 'true' : 'false');
+  }
+  // Redirect using NodeJS HTTP
+  res.writeHead(302, { Location: redirectUrl.toString() });
+  res.end();
+}
+
+module.exports = redirectToFront;
